refactored initalize command and formatted with black and isort
