+++ automation/Commands/command_executor.py
@@ -1,7 +1,6 @@
 from ..Errors import CommandExecutionError
 from . import browser_commands, profile_commands
-from .types import (BaseCommand, DumpProfCommand,
-                    InitializeCommand, RunCustomFunctionCommand)
+from .types import BaseCommand, DumpProfCommand, RunCustomFunctionCommand
 
 
 def execute_command(
@@ -51,11 +50,6 @@ def execute_command(
             "extension_socket": extension_socket,
         }
         command.function_handle(*command.func_args, **arg_dict)
-
-    elif type(command) is InitializeCommand:
-        browser_commands.initialize(
-            visit_id=command.visit_id, extension_socket=extension_socket
-        )
 
     else:
         raise CommandExecutionError("Invalid Command", command)
